requirements.txt
streamlit
requests
fastapi
uvicorn
openai
pdfplumber
python-dotenv
python-multipart